issue - 1 : moved validator and catches from service to controller
