+++ controllers/blog.controller.js
@@ -1,5 +1,6 @@
 const blogService = require("../services/blog.service");
 const { sendResponse } = require("../utils/contentNegotiation");
+const { AppError } = require("../utils/errorHandler");
 
 ("use strict");
 
@@ -17,6 +18,10 @@ exports.getAllBlogs = async (req, res, next) => {
 
 exports.createBlog = async (req, res, next) => {
   try {
+    const {title, description } = req.body;
+    if (!title || !description) {
+      throw new AppError("Title and description are needed", 400, false);
+    }
     const createdBlog = await blogService.createBlog(req.body, req.username);
     sendResponse(req, res, 201, createdBlog);
   } catch (err) {
@@ -35,6 +40,10 @@ exports.getBlogById = async (req, res, next) => {
 
 exports.editBlogByBlogId = async (req, res, next) => {
   try {
+    const blogItemsToEdit = req.body;
+    if (!blogItemsToEdit.title && !blogItemsToEdit.description) {
+      throw new AppError("Title and description are missing", 400, false);
+    }
     const editedBlog = await blogService.editBlogByBlogId(req.params.blogId, req.body);
     sendResponse(req, res, 200, "Blog edited successfully");
   } catch (err) {
