+++ services/blog.service.js
@@ -2,14 +2,11 @@ const blogRepository = require("../repositories/blog.repository");
 const { AppError } = require("../utils/errorHandler");
 const userService = require("../services/user.service");
 const { setLimitAndOffset } = require("../utils/pagination");
+const BlogDTO = require('../DTOs/blog.dto');
 
 ("use strict");
 
 async function createBlog(blog, username) {
-  if (!blog.title || !blog.description) {
-    throw new AppError("Title and description are needed", 400, false);
-  }
-
     const authorExists = await userService.getUserByUsername(username);
     if(!authorExists){
       throw new AppError("Author does not exist", 404, false);
@@ -20,53 +17,36 @@ async function createBlog(blog, username) {
 }
 
 async function getAllBlogs(pageSize, pageNumber) {
-  try {
     const { limit , offset } = setLimitAndOffset(pageSize, pageNumber);
     const data = await blogRepository.getAllBlogs(limit, offset);
-    return data;
-  } catch {
-    throw new AppError("Cannot find any Blogs table", 404, false);
-  }
+    const allBlog = [];
+    data.forEach((element) => {
+      allBlog.push( new BlogDTO(element));
+    });
+    return allBlog;
 }
 
 async function editBlogByBlogId(blogId, blogItemsToEdit) {
-
-  try {
-    if (!blogItemsToEdit.title && !blogItemsToEdit.description) {
-      throw new AppError("Title and description are missing", 400, false);
-    }
     const result = await blogRepository.editBlogByBlogId(blogId, blogItemsToEdit);
     return result;
-  } catch (err) {
-    throw new AppError( err.message, err.statusCode, err.isOperational);
-  }
 }
 
 async function deleteBlogByBlogId(blogId) {
-  try {
     const result = await blogRepository.deleteBlogByBlogId(blogId);
+    if(!result) throw new AppError("Blog not found", 404, false);
     return result;
-  } catch {
-    throw new AppError("Blog not found", 404, false);
-  }
 }
 
 async function getBlogById(blogId) {
-  try {
+  
     const result = await blogRepository.getBlogById(blogId);
-    return result;
-  } catch {
-    throw new AppError("Blog not found", 404, false);
-  }
+    if(!result) throw new AppError("Blog not found", 404, false);
+    return new BlogDTO(result); 
 }
 
 async function getBlogByAuthorId(authorId) {
-  try {
     const blogExists = await blogRepository.getBlogByAuthorId(authorId);
     return blogExists;
-  } catch {
-    throw new AppError("Blog not found", 404, false);
-  }
 }
 
 module.exports = {
